% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amle_fun.R
\name{amle_fun}
\alias{amle_fun}
\title{Rasch Anchored Maximum Likelihood Estimation (AMLE) Function with know item difficulties}
\usage{
amle_fun(response, response_max, scale_id, scale_cuts, it_diffic)
}
\arguments{
\item{response}{is a vector of scored responses.}

\item{response_max}{is a vector of the maximum of the scored responses.}

\item{scale_id}{is a vector indicating the scale for each item. The scale ID must be the row ID in 'scale_cuts'.}

\item{scale_cuts}{is a list of the cuts/thresholds for each scale type. Should be ordered to ID with 'scale_id'.}

\item{it_diffic}{is a vector of item difficulties.}
}
\description{
Estimates the abilities based on known item difficulties. More info here: https://www.rasch.org/rmt/rmt122q.htm
}
\examples{
set.seed(1010)
#Simulate some response string
resp = c(1,1,2,2,3,3,4,4)

#Simulate maximum responses
max_responses = c(2,2,3,3,3,3,4,4)

#Identify the scale type for each item
scale_ID = c(1,1,2,2,2,2,3,3)

#List of a made-up scale thresholds
final_scale = list("ScaleA" = c(-0.2,0.2),
                   "ScaleB" = c(-0.3, -0.2, 0.5),
                   "ScaleC" = c(-0.4, -0.1, 0.3, 0.2))

#Simulate item difficulty parameters
item_difficulties = c(runif(8, min = -3, max = 3) )

rasch_amlestimate(resp, max_responses, scale_ID, final_scale, item_difficulties)
}
\keyword{AMLE}
\keyword{Muti-stage}
\keyword{Rasch}
\keyword{adaptive}
\keyword{estimation}
\keyword{mle}
