% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_respgen_fun.R
\name{rasch_respgen_fun}
\alias{rasch_respgen_fun}
\title{Rasch partial credit model response generating function}
\usage{
rasch_respgen_fun(pers_ability, item_diff, item_thrsh)
}
\arguments{
\item{pers_ability}{is a value equal to the personal ability measure}

\item{item_diff}{is a value equal to the item difficulty.}

\item{item_thrsh}{is a vector of the item thresholds. If dichotomous, item_thrsh will be a single value.}
}
\description{
Creates responses based on the Rasch partial credit model.
}
\examples{
set.seed(1010)
ability = rnorm(1,0,1)  #simulated person ability
difficulty  = rnorm(1,0,1)  #simulated item difficulty

a = rnorm(2, 0, 1)
b = -a
thresholds = c(a, b) #simulated threshold measures

rasch_respgen_fun(ability, difficulty, thresholds)
}
\keyword{credit}
\keyword{generation}
\keyword{mcmc}
\keyword{model}
\keyword{partial}
\keyword{pcm}
\keyword{response}
\keyword{simulation}
