% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respgen_fun.R
\name{respgen_fun}
\alias{respgen_fun}
\title{Rasch model response generating function}
\usage{
respgen_fun(theta, diff, mscale, scale_group)
}
\arguments{
\item{theta}{is a vector of scored responses.}

\item{diff}{is a vector of the maximum of the scored responses.}

\item{mscale}{is a vector indicating the scale for each item. The scale ID must be the row ID in 'scale_cuts'.}

\item{scale_group}{is a list of the cuts/thresholds for each scale type. Should be ordered to ID with 'scale_id'.}
}
\description{
Creates responses based on the Rasch model.
}
\examples{
set.seed(1010)
}
\keyword{generation}
\keyword{mcmc}
\keyword{response}
\keyword{simulation}
